feat: 기본 navbar notification icon 추가
