+++ src/pages/Navbar/Navbar.js
@@ -1,43 +1,120 @@
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import Button from '@mui/material/Button';
+import AdbIcon from '@mui/icons-material/Adb';
+import NotificationsIcon from '@mui/icons-material/Notifications';
+import { useState } from 'react';
 import { useRecoilState } from 'recoil';
-import { userAtom } from '../../states';
-import { Link, useNavigate } from 'react-router-dom';
-import toast from 'react-hot-toast';
+import { userAtom } from 'states';
+import { AnchorNav, UserRelatedDropDown } from 'components/Navbar';
+import { Link } from 'react-router-dom';
 
-const Navbar = () => {
-  const [user, setUser] = useRecoilState(userAtom);
-  const navigate = useNavigate();
+const ResponsiveAppBar = () => {
+  const [user] = useRecoilState(userAtom);
+  const [anchorElNav, setAnchorElNav] = useState(null);
+  const [anchorElUser, setAnchorElUser] = useState(null);
 
-  const logout = async () => {
-    toast.success('로그아웃 성공 👍');
-    await setUser(null);
-    await localStorage.clear();
-    navigate('/');
+  const handleOpenNavMenu = (event) => {
+    setAnchorElNav(event.currentTarget);
+  };
+  const handleOpenUserMenu = (event) => {
+    setAnchorElUser(event.currentTarget);
+  };
+
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
   };
 
   return (
-    <div>
-      <div>Nav</div>
-      <Link to="/">
-        <button>Home</button>
-      </Link>
-      {user && (
-        <>
-          <div>Welcome {user.username}!</div>
-          <button onClick={logout}>Logout</button>
-        </>
-      )}
-      {!user && (
-        <>
-          <Link to="/login">
-            <button>Login</button>
+    <AppBar position="static" color="primary">
+      <Container maxWidth="xl">
+        <Toolbar disableGutters>
+          <AdbIcon sx={{ display: { xs: 'none', md: 'flex' }, mr: 1 }} />
+          <Link to="/">
+            <Typography
+              variant="h6"
+              noWrap
+              component="a"
+              sx={{
+                mr: 2,
+                display: { xs: 'none', md: 'flex' },
+                fontFamily: 'monospace',
+                fontWeight: 700,
+                letterSpacing: '.3rem',
+                color: 'inherit',
+                textDecoration: 'none',
+              }}
+            >
+              SELLON
+            </Typography>
           </Link>
-          <Link to="/register">
-            <button>Signup</button>
-          </Link>
-        </>
-      )}
-    </div>
+          <AnchorNav
+            anchorElNav={anchorElNav}
+            openNavMenu={handleOpenNavMenu}
+            closeNavMenu={handleCloseNavMenu}
+          />
+          <AdbIcon sx={{ display: { xs: 'flex', md: 'none' }, mr: 1 }} />
+          <Typography
+            variant="h5"
+            noWrap
+            component="a"
+            sx={{
+              mr: 2,
+              display: { xs: 'flex', md: 'none' },
+              flexGrow: 1,
+              fontFamily: 'monospace',
+              fontWeight: 700,
+              letterSpacing: '.3rem',
+              color: 'inherit',
+              textDecoration: 'none',
+            }}
+          >
+            <Link to="/">SELLON</Link>
+          </Typography>
+          <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
+            <Button
+              onClick={handleCloseNavMenu}
+              sx={{ my: 2, color: 'white', display: 'block' }}
+            >
+              소개페이지
+            </Button>
+            <Button
+              onClick={handleCloseNavMenu}
+              sx={{ my: 2, color: 'white', display: 'block' }}
+            >
+              경매장
+            </Button>
+            {!user && (
+              <Link to="/login">
+                <Button
+                  onClick={handleCloseNavMenu}
+                  sx={{ my: 2, color: 'white', display: 'block' }}
+                >
+                  로그인/회원가입
+                </Button>
+              </Link>
+            )}
+          </Box>
+          {user && (
+            <>
+              <NotificationsIcon sx={{ mr: '2rem' }} />
+              <UserRelatedDropDown
+                openUserMenu={handleOpenUserMenu}
+                closeUserMenu={handleCloseUserMenu}
+                anchorElUser={anchorElUser}
+              />
+            </>
+          )}
+        </Toolbar>
+      </Container>
+    </AppBar>
   );
 };
-
-export default Navbar;
+export default ResponsiveAppBar;
