+++ src/pages/index.js
@@ -2,4 +2,3 @@ export { default as Home } from 'pages/Home/Home';
 export { default as Login } from 'pages/Login/Login';
 export { default as Register } from 'pages/Register/Register';
 export { default as Navbar } from 'pages/Navbar/Navbar';
-export { default as NavbarTest } from 'pages/Navbar/NavbarTest';
